+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Pass correct exit return codes (bug #18)
+- Autocommit is explicitly turned off (bug #22)
 
 ## [1.3.1] - 2019-07-28